Added ValidMoves state and css class indicator
Added a 2d array to hold valid moves for the picked up piece. Using a
simple circle as the place holder for squares that are a valid move.

Right now we're just setting the (3,3) position for any piece that is
picked up. Will work on generating actual valid moves next.